backend/requirements.txt
asgiref==3.8.1
bleach==6.1.0
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
webencodings==0.5.1
whitenoise==6.6.0
